Add image for bookstack
We are planning to replace the current wikimedia for a bookstack app,
which is more modern and easier to manage.

The docker compose settings are based on
https://github.com/solidnerd/docker-bookstack/blob/master/docker-compose.yml

We are loading the DB credentials from a .env file which we will set
manually on the server. Also, the "/wiki2" URL is temporary while we
keep the old wiki around.

Signed-off-by: @pedrotrschneider